#511 - Move ValidateXML and ValidateXMLs annotation to coffee-dto-base
