+++ coffee-dto/coffee-dto-base/pom.xml
@@ -10,7 +10,7 @@
 
 	<properties>
 		<version.jakarta.xml.bind.jaxb-api>4.0.0</version.jakarta.xml.bind.jaxb-api>
-
+		<version.jakarta.enterprise.cdi-api>4.0.1</version.jakarta.enterprise.cdi-api>
 		<!-- Test dependency versions -->
 		<version.junit.jupiter>5.9.1</version.junit.jupiter>
 		<version.org.mockito>4.9.0</version.org.mockito>
@@ -21,6 +21,11 @@
 			<groupId>jakarta.xml.bind</groupId>
 			<artifactId>jakarta.xml.bind-api</artifactId>
 			<version>${version.jakarta.xml.bind.jaxb-api}</version>
+		</dependency>
+		<dependency>
+			<groupId>jakarta.enterprise</groupId>
+			<artifactId>jakarta.enterprise.cdi-api</artifactId>
+			<version>${version.jakarta.enterprise.cdi-api}</version>
 		</dependency>
 		<dependency>
 			<groupId>hu.icellmobilsoft.coffee</groupId>
