+++ coffee-dto/coffee-dto-base/src/main/java/hu/icellmobilsoft/coffee/dto/annotation/Version.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package hu.icellmobilsoft.coffee.cdi.annotation;
+package hu.icellmobilsoft.coffee.dto.annotation;
 
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
