+++ coffee-dto/coffee-dto-base/src/main/java/hu/icellmobilsoft/coffee/dto/annotation/validation/xml/ValidateXMLs.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package hu.icellmobilsoft.coffee.rest.validation.xml.annotation;
+package hu.icellmobilsoft.coffee.dto.annotation.validation.xml;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -26,35 +26,36 @@ import java.lang.annotation.Target;
 
 import jakarta.enterprise.util.Nonbinding;
 
-import hu.icellmobilsoft.coffee.cdi.annotation.Version;
-
 /**
  * Ha egy RestService-ben szereplő metódus a request body paraméterét megannotáljuk ezzel az annotációval, akkor a deszerializációt és a validációt a
  * MessageBodyReaderBase megfelelő implementáció végzik
  *
- * @see ValidateXMLs
- * @see Version
- * @author attila.nyers
+ * Példa a használatra:
+ *
+ * <pre>
+ * ExampleResponse postExampleRequest(@Context HttpHeaders headers, @Context HttpServletRequest servletRequest,
+ *         &#64;ValidateXMLs({ @ValidateXML(version = @Version(include = @Range(from = "1.0", to = "1.9")), xsdPath = ""), // Nincs XSD validáció
+ *                 &#64;ValidateXML(version = @Version(include = @Range(from = "1.10")), xsdPath = "sample.xsd") }) ExampleRequest exampleRequest)
+ *         throws BaseException;
+ * </pre>
+ *
+ * Lehet egy ValidateXML annotáció is önmagában.
+ *
+ * Teljes használati leírás: /docs/howto/xsd_xml_validation_depend_on_version.adoc
+ *
+ * @see ValidateXML
  * @author ferenc.lutischan
  * @since 1.0.0
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.PARAMETER })
-public @interface ValidateXML {
+public @interface ValidateXMLs {
 
     /**
-     * Melyik verziókra aktiválódjon
+     * Az egyes xsd validációk verziófüggő felsorolása (azaz Verzió -&gt; xsd összerendelés)
      *
-     * @return {@link Version}
+     * @return {@code ValidateXML} tömb
      */
     @Nonbinding
-    Version version() default @Version();
-
-    /**
-     * A Version-höz rendelt xsd fájl elérési útvonallal
-     *
-     * @return path of XSD file
-     */
-    @Nonbinding
-    String xsdPath();
+    ValidateXML[] value();
 }
